rename files, refine devbar
