+++ pages/UserDashboard.tsx
@@ -1,38 +1,34 @@
 import Link from 'next/link';
-import { GridBackgroundDemo } from "../components/GridBackground";
+import { GridBackgroundDemo } from '../components/GridBackground';
 import { InfiniteMovingCards } from '../components/Infinite-Moving-Cards';
 
-import NavBar from "../components/NavBar";
-import Footer from "../components/Footer";
+import NavBar from '../components/NavBar';
+import Footer from '../components/Footer';
 
 const userDashboardItems = [
   {
-    quote: "Not that expensive.",
-    name: "Cluster Cost",
-    title: "Daily Overview"
+    quote: 'Not that expensive.',
+    name: 'Cluster Cost',
+    title: 'Daily Overview',
   },
   {
-    quote: "But this week was relatively expensive.",
-    name: "Cluster Cost",
-    title: "Weekly Overview"
+    quote: 'But this week was relatively expensive.',
+    name: 'Cluster Cost',
+    title: 'Weekly Overview',
   },
   {
-    quote: "Still within budget for the month though!",
-    name: "Cluster Cost",
-    title: "Monthly Overview"
+    quote: 'Still within budget for the month though!',
+    name: 'Cluster Cost',
+    title: 'Monthly Overview',
   },
 ];
 
 const UserDashboard = () => {
   return (
-    <div className="relative h-[50rem]">
-      {" "}
-      {/* Set the relative positioning and height */}
+    <div className='relative h-[50rem]'>
       <NavBar />
       <GridBackgroundDemo />
-      <div className="absolute top-[100px] inset-0 flex items-center justify-center">
-        {" "}
-        {/* Center the content */}
+      <div className='absolute top-[100px] inset-0 flex items-center justify-center'>
         <InfiniteMovingCards items={userDashboardItems} />
       </div>
       <Footer />
@@ -40,6 +36,4 @@ const UserDashboard = () => {
   );
 };
 
-
 export default UserDashboard;
-
